+++ xcube/core/tile.py
@@ -29,7 +29,8 @@ from deprecated import deprecated
 from xcube.core.schema import get_dataset_xy_var_names
 # Exported for backward compatibility only
 # noinspection PyUnresolvedReferences
-from ._tile2 import (compute_rgba_tile,
+from ._tile2 import (compute_tiles,
+                     compute_rgba_tile,
                      get_var_valid_range,
                      get_var_cmap_params,
                      DEFAULT_VALUE_RANGE,